fix(edit): Redefine custom chart definition on sources when editing variables
Editing a variable will trigger a source update to ensure a valid configuration is set for sources
that have the custom chart enabled. The default chart configuration set by the admins will be
retained for the variables that were not modified and updated/new variables will get a new default
configuration.